currency interpretation
tableConvert uses the general addRecord function which now uses the
o.set function (rather than editing the value array directly).

This then allows live interpretation of data around currency - leading
currency markers ($,£ etc) are stripped from data based on a
possiblyCurrency array in the graph settings, and then currency is set
in the graph object itself - later this can be used as a marker on the y
axis for example.