change of theme loader
used to use jQuery's getScript, but that caused problems with code
ordering (runs in a separate thread) - is possible to force the
attribute loading after file has loaded, but by that time animation may
have started and it adjusts as it's drawing, not a good look.

So now the html has to have script tags for the additional theme files.